Time: 136 ms (54.93%) | Memory: 44.8 MB (47.89%) - LeetSync

+++ 3794-find-the-minimum-amount-of-time-to-brew-potions/find-the-minimum-amount-of-time-to-brew-potions.java
@@ -4,10 +4,10 @@ class Solution {
         long[] done = new long[n + 1];
          for (int j = 0; j < m; ++j) {
             for (int i = 0; i < n; ++i) {
-                done[i + 1] = Math.max(done[i + 1], done[i]) + (long) mana[j] * skill[i];
+                done[i + 1] = Math.max(done[i + 1], done[i]) + mana[j] * skill[i];
             }
             for (int i = n - 1; i > 0; --i) {
-                done[i] = done[i + 1] -  (long) mana[j] * skill[i];
+                done[i] = done[i + 1]-mana[j] * skill[i];
             }
         }
